pyproject.toml
[build-system]
requires = ["setuptools >= 65", "wheel"]
build_backend = "setuptools.build_meta"

[project]
name = "root-kit"
version = "0.0.1"
authors = [
  { name="Basti", email="enrm000@gmail.com" },
]
description = "CLI Tool for installing Root-Kit Application Code"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "typer",
    "gitpython",
    "python-dotenv",
    "docker"
]

[project.urls]
CLI = "https://github.com/batic420/CLI-Tool.git"
Application = "https://github.com/batic420/Root-Kit.git"

[project.scripts]
root-kit = "root_kit.cli:rt"